Create RobTheHouseDP.cpp
The "Rob the house" problem is a classic problem in dynamic programming and algorithmic optimization.
We are planning to rob houses along a street. Each house has a certain amount of money stashed. However, adjacent houses have security systems connected, and if two adjacent houses are robbed, the security system will immediately alert the police.