TPrimeChecker.cpp
The TPrimeChecker program checks whether numbers in an array are T-prime or not. It defines two functions, isPrime and isTPrime, to determine if a number is prime and if its square root is prime, respectively. The main function iterates through an array of integers and prints whether each number is T-prime or not.